Use custom app json encoder - remove hacky API class in favor of flask-yoloapi
